lipsync_processor.py
import cv2
import numpy as np
import tensorflow as tf
import face_recognition
from moviepy.editor import VideoFileClip, AudioFileClip
import os
import tempfile
from PIL import Image

class LipSyncProcessor:
    def __init__(self):
        # Carregar o modelo Wav2Lip (você precisará baixar os pesos do modelo)
        self.model = self._load_model()
        
    def _load_model(self):
        # Implementar carregamento do modelo Wav2Lip
        # Retornar o modelo carregado
        pass
        
    def _extract_audio_features(self, audio_path):
        """Extrai características do áudio para o modelo Wav2Lip."""
        # Implementar extração de características do áudio
        pass

    def _get_face_region(self, frame, face_location):
        """Extrai a região do rosto da imagem."""
        top, right, bottom, left = face_location
        return frame[top:bottom, left:right]

    def _apply_lipsync(self, face_region, audio_features):
        """Aplica o lipsync em uma região do rosto."""
        # Implementar a lógica do Wav2Lip
        pass

    def _blend_face(self, original_frame, new_face_region, face_location):
        """Mistura o rosto processado de volta no frame original."""
        top, right, bottom, left = face_location
        result = original_frame.copy()
        result[top:bottom, left:right] = new_face_region
        return result

    def process_video(self, video_path, audio_path, face_id=0):
        """Processa um vídeo com lipsync."""
        video = VideoFileClip(video_path)
        audio = AudioFileClip(audio_path)
        
        # Extrair características do áudio
        audio_features = self._extract_audio_features(audio_path)
        
        # Criar arquivo temporário para o resultado
        with tempfile.NamedTemporaryFile(suffix='.mp4', delete=False) as temp_output:
            output_path = temp_output.name
        
        # Processar cada frame
        frames = []
        for frame in video.iter_frames():
            # Detectar rostos
            face_locations = face_recognition.face_locations(frame)
            
            if len(face_locations) > face_id:
                face_location = face_locations[face_id]
                face_region = self._get_face_region(frame, face_location)
                
                # Aplicar lipsync
                new_face_region = self._apply_lipsync(face_region, audio_features)
                
                # Misturar o rosto processado de volta no frame
                processed_frame = self._blend_face(frame, new_face_region, face_location)
                frames.append(processed_frame)
            else:
                frames.append(frame)
        
        # Criar vídeo final
        output_video = VideoFileClip(frames, fps=video.fps)
        final_video = output_video.set_audio(audio)
        final_video.write_videofile(output_path)
        
        return output_path

    def process_image(self, image_path, audio_path, face_id=0):
        """Processa uma imagem estática com lipsync."""
        # Carregar imagem
        image = cv2.imread(image_path)
        image_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        
        # Detectar rostos
        face_locations = face_recognition.face_locations(image_rgb)
        
        if len(face_locations) <= face_id:
            raise ValueError("Face ID não encontrado na imagem")
        
        face_location = face_locations[face_id]
        face_region = self._get_face_region(image_rgb, face_location)
        
        # Extrair características do áudio
        audio_features = self._extract_audio_features(audio_path)
        
        # Criar frames para o vídeo
        frames = []
        for _ in range(int(AudioFileClip(audio_path).duration * 30)):  # 30 fps
            new_face_region = self._apply_lipsync(face_region, audio_features)
            processed_frame = self._blend_face(image_rgb, new_face_region, face_location)
            frames.append(processed_frame)
        
        # Criar arquivo temporário para o resultado
        with tempfile.NamedTemporaryFile(suffix='.mp4', delete=False) as temp_output:
            output_path = temp_output.name
        
        # Criar vídeo final
        output_video = VideoFileClip(frames, fps=30)
        audio = AudioFileClip(audio_path)
        final_video = output_video.set_audio(audio)
        final_video.write_videofile(output_path)
        
        return output_path

    def process_media(self, media_path, audio_path, face_id=0):
        """Processa mídia (vídeo ou imagem) com lipsync."""
        if media_path.lower().endswith(('.mp4')):
            return self.process_video(media_path, audio_path, face_id)
        elif media_path.lower().endswith(('.png', '.jpg', '.jpeg')):
            return self.process_image(media_path, audio_path, face_id)
        else:
            raise ValueError("Formato de arquivo não suportado")
