requirements.txt
flask>=2.3.3
opencv-python>=4.8.1.78
numpy>=1.24.3
tensorflow>=2.13.0
librosa>=0.10.1
scipy>=1.11.3
face-recognition>=1.3.0
moviepy>=1.0.3
pillow>=10.0.1
dlib>=19.24.2
